% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluateForecast}
\alias{evaluateForecast}
\title{Graphically evaluate a forecast}
\usage{
evaluateForecast(play_by_play)
}
\arguments{
\item{play_by_play}{The dataset containing actual and predicted values for an
event}
}
\value{

}
\description{
Plots the predicted vs. actual values to visually see how well
  the model performed.
}
