% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Evaluate matchup models}
\usage{
evaluateModel(train_data, test_data, x, y, z, model)
}
\arguments{
\item{train_data}{The data to train the model on}

\item{test_data}{The data ot test the model on}

\item{x}{The column name for the first opponent (e.g. pitcher)}

\item{y}{The column name for the second opponent (e.g. batter)}

\item{z}{The colunn name for the global average (e.g. league)}

\item{model}{A model object that will be evaluated}
}
\value{
A plot showing how well predictions match actual values, and a Brier Score which is a numerical value that can be used to compare across various models
}
\description{
Evaluate matchup models
}
