% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glmPrediction}
\alias{glmPrediction}
\title{Eevent prediction using logistic regression}
\usage{
glmPrediction(train_data, test_data, x = x, y = y, z = z)
}
\arguments{
\item{train_data}{The set of data to train the emodel on}

\item{test_data}{The set of data to caluclate predicted outcomes}

\item{x}{The name of the column for the first opponent (e.g. pitcher)}

\item{y}{The name of the column for the second opponent (e.g. batter)}

\item{z}{The name of the column for the population average (e.g. league)}
}
\value{
A result containing the model that was derived, the predicted
  outcomes, and the coefficients for the model (useful for comparing to log5
  model which has coefficientts (1,1,1))
}
\description{
Use logistic regression to calculate the predicted outcome of an
  event between two opponents given their individual rates for that event
}
