% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{groupAvg}
\alias{groupAvg}
\title{Average by group}
\usage{
groupAvg(play_by_play, group)
}
\arguments{
\item{play_by_play}{the data frame to get averages (means) from}

\item{group}{the columns to group together for the average (mean)}
}
\description{
Average by group
}
