% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{log5Prediction}
\alias{log5Prediction}
\title{The log5 method}
\usage{
log5Prediction(train_data, test_data, c1, c2, c3)
}
\arguments{
\item{train_data}{not used in this function, but required for compatibility
with generic models}

\item{test_data}{the dataset containing the required 3 columns}

\item{c1}{name of column in test_data containing rates for the first opponent
(typically the pitcher)}

\item{c2}{name of column in test_data containing rates for the second
opponent (typically the batter)}

\item{c3}{name of column in test_data containing rates for the population
(typicallly the league)}
}
\value{
a dataset consisting of the predicted outcome based on the log5
  formula and coefficients which are all 1 for the log 5 model
}
\description{
This calculates the predicted outcome between two opponents
  based on their individual rates for that event. Originally derived by Bill
  James
}
\examples{
log5Prediction(NULL, data.frame('pitcher'=.200,
                                          'batter'=.300,
                                          'league'=.250),
                         'pitcher', 'batter', 'league')
}
