% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{previousYearForecast}
\alias{previousYearForecast}
\title{Add forecast using previous year's average}
\usage{
previousYearForecast(play_by_play, group)
}
\arguments{
\item{play_by_play}{The dataset containing event data across the league}

\item{group}{The group by which averages should be calculated}
}
\value{
The previous year's averages, which will be used as the forecast for
  this year's averages
}
\description{
Add forecast using previous year's average
}
