% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{printBrierSummary}
\alias{printBrierSummary}
\title{Print summary of Brier Scores}
\usage{
printBrierSummary(model_list, bs_ref)
}
\arguments{
\item{model_list}{A list of objects with a $brierscore attribute}

\item{bs_ref}{The reference brier score}
}
\description{
Print summary of Brier Scores
}
