% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{printCoefSummary}
\alias{printCoefSummary}
\title{Print summary of coefficients}
\usage{
printCoefSummary(model_list)
}
\arguments{
\item{model_list}{A list of objects with $coefficieints attribute}
}
\description{
Print summary of coefficients
}
